+++ .idea/libraries/appcompat_v7_25_3_0.xml
@@ -1,12 +1,12 @@
 <component name="libraryTable">
   <library name="appcompat-v7-25.3.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/2da4de2358b30baaef19bad877f9379c513c4381/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/2da4de2358b30baaef19bad877f9379c513c4381/output/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/4b7b3b3b6787c3f33e8543f45656b9603134e32d/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/4b7b3b3b6787c3f33e8543f45656b9603134e32d/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
-      <root url="jar://$PROJECT_DIR$/../../Android/sdk/extras/android/m2repository/com/android/support/appcompat-v7/25.3.0/appcompat-v7-25.3.0-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../sdk/extras/android/m2repository/com/android/support/appcompat-v7/25.3.0/appcompat-v7-25.3.0-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
