+++ .idea/libraries/constraint_layout_1_0_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="constraint-layout-1.0.2">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/8bc616c5348727e1345b82fd455b8e9b45627bd8/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/8bc616c5348727e1345b82fd455b8e9b45627bd8/output/res" />
+      <root url="file://$USER_HOME$/.android/build-cache/ac887715cf5710a557e1b049bf8d72b7f3fc16e4/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/ac887715cf5710a557e1b049bf8d72b7f3fc16e4/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
