+++ .idea/libraries/kotlin_stdlib_1_0_0.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="kotlin-stdlib-1.0.0">
+    <CLASSES>
+      <root url="jar://$APPLICATION_HOME_DIR$/gradle/m2repository/org/jetbrains/kotlin/kotlin-stdlib/1.0.0/kotlin-stdlib-1.0.0.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
