+++ .idea/libraries/support_compat_25_3_0.xml
@@ -1,12 +1,12 @@
 <component name="libraryTable">
   <library name="support-compat-25.3.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/9bc5142b0d586b7caf8b7609303f8357750f1f7c/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/9bc5142b0d586b7caf8b7609303f8357750f1f7c/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/f445e4a44adc237dd6e62730923b0743c1fa7d59/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/f445e4a44adc237dd6e62730923b0743c1fa7d59/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
-      <root url="jar://$PROJECT_DIR$/../../Android/sdk/extras/android/m2repository/com/android/support/support-compat/25.3.0/support-compat-25.3.0-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../sdk/extras/android/m2repository/com/android/support/support-compat/25.3.0/support-compat-25.3.0-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
