+++ .idea/libraries/support_core_ui_25_3_0.xml
@@ -1,12 +1,12 @@
 <component name="libraryTable">
   <library name="support-core-ui-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/fec4705a7b9f920bc84a5baddb8c0f4f614ccee7/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/fec4705a7b9f920bc84a5baddb8c0f4f614ccee7/output/res" />
+      <root url="file://$USER_HOME$/.android/build-cache/eba1a336cc0090add361cc54c361c41a47b5ba05/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/eba1a336cc0090add361cc54c361c41a47b5ba05/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
-      <root url="jar://$PROJECT_DIR$/../../Android/sdk/extras/android/m2repository/com/android/support/support-core-ui/25.3.0/support-core-ui-25.3.0-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../sdk/extras/android/m2repository/com/android/support/support-core-ui/25.3.0/support-core-ui-25.3.0-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
