+++ .idea/libraries/support_core_utils_25_3_0.xml
@@ -1,12 +1,12 @@
 <component name="libraryTable">
   <library name="support-core-utils-25.3.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/6997488f89c837e83ee1890ef2875cea5907cf9c/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/6997488f89c837e83ee1890ef2875cea5907cf9c/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/c942d05520b44ee2aea1295f7dae4f36453b050d/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/c942d05520b44ee2aea1295f7dae4f36453b050d/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
-      <root url="jar://$PROJECT_DIR$/../../Android/sdk/extras/android/m2repository/com/android/support/support-core-utils/25.3.0/support-core-utils-25.3.0-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../sdk/extras/android/m2repository/com/android/support/support-core-utils/25.3.0/support-core-utils-25.3.0-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
