+++ .idea/libraries/support_fragment_25_3_0.xml
@@ -1,12 +1,12 @@
 <component name="libraryTable">
   <library name="support-fragment-25.3.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/3f76bad6c25091fd5092c228cb00a17083897d79/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/3f76bad6c25091fd5092c228cb00a17083897d79/output/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/309ade0e2d1af6f986a9c98d74420c42091b1a6e/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/309ade0e2d1af6f986a9c98d74420c42091b1a6e/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
-      <root url="jar://$PROJECT_DIR$/../../Android/sdk/extras/android/m2repository/com/android/support/support-fragment/25.3.0/support-fragment-25.3.0-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../sdk/extras/android/m2repository/com/android/support/support-fragment/25.3.0/support-fragment-25.3.0-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
