+++ .idea/libraries/support_media_compat_25_3_0.xml
@@ -1,12 +1,12 @@
 <component name="libraryTable">
   <library name="support-media-compat-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/cae1e14a5057d4b21583d9568a80f2d89e75420c/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/cae1e14a5057d4b21583d9568a80f2d89e75420c/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/8e5e4e61696a3cd34d4f13b2a819d109c80b6141/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/8e5e4e61696a3cd34d4f13b2a819d109c80b6141/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
-      <root url="jar://$PROJECT_DIR$/../../Android/sdk/extras/android/m2repository/com/android/support/support-media-compat/25.3.0/support-media-compat-25.3.0-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../sdk/extras/android/m2repository/com/android/support/support-media-compat/25.3.0/support-media-compat-25.3.0-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
