+++ .idea/libraries/support_v4_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="support-v4-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/ea0d5034b10c4ba23f3cb8ced70216776ebd28d6/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/ea0d5034b10c4ba23f3cb8ced70216776ebd28d6/output/res" />
+      <root url="file://$USER_HOME$/.android/build-cache/77dd393a15ece3ff54234eb389f491161d87b70e/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/77dd393a15ece3ff54234eb389f491161d87b70e/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
