+++ .idea/libraries/support_vector_drawable_25_3_0.xml
@@ -1,12 +1,12 @@
 <component name="libraryTable">
   <library name="support-vector-drawable-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/23390ae60415381f6f4057a791d76074b8616a82/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/23390ae60415381f6f4057a791d76074b8616a82/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/163bbb4b8363596056ba8ef5475afe15d233f943/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/163bbb4b8363596056ba8ef5475afe15d233f943/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
-      <root url="jar://$PROJECT_DIR$/../../Android/sdk/extras/android/m2repository/com/android/support/support-vector-drawable/25.3.0/support-vector-drawable-25.3.0-sources.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../sdk/extras/android/m2repository/com/android/support/support-vector-drawable/25.3.0/support-vector-drawable-25.3.0-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
