+++ fourfoldloader/src/main/java/com/agrawalsuneet/fourfoldloader/basicviews/LoaderContract.kt
@@ -1,0 +1,10 @@
+package com.agrawalsuneet.fourfoldloader.basicviews
+
+import android.util.AttributeSet
+
+/**
+ * Created by suneet on 10/20/17.
+ */
+interface LoaderContract {
+    fun initAttributes(attrs: AttributeSet)
+}
